+++ src/roiextractors/extractors/schnitzerextractor/cnmfesegmentationextractor.py
@@ -138,9 +138,5 @@ class CnmfeSegmentationExtractor(SegmentationExtractor):
             if segmentation_object.get_sampling_frequency() is not None:
                 inputoptions.create_dataset("Fs", data=segmentation_object.get_sampling_frequency())
 
-    # defining the abstract class enformed methods:
-    def get_roi_ids(self):
-        return list(range(self.get_num_rois()))
-
     def get_image_size(self):
         return self._image_masks.shape[0:2]