+++ tests/test_internals/test_testing_tools.py
@@ -2,6 +2,9 @@ import numpy as np
 from hdmf.testing import TestCase
 from numpy.testing import assert_array_equal
 
+import numpy as np
+from numpy.testing import assert_array_equal
+
 from roiextractors.testing import (
     generate_dummy_segmentation_extractor,
     _assert_iterable_complete,
@@ -33,6 +36,11 @@ class TestDummySegmentationExtractor(TestCase):
         assert segmentation_extractor.get_accepted_list() == segmentation_extractor.get_roi_ids()
         assert segmentation_extractor.get_rejected_list() == []
         assert segmentation_extractor.get_roi_locations().shape == (2, self.num_rois)
+
+        # Test frame_to_time
+        times = np.round(np.arange(self.num_frames) / self.sampling_frequency, 6)
+        assert_array_equal(segmentation_extractor.frame_to_time(frame_indices=np.arange(self.num_frames)), times)
+        self.assertEqual(segmentation_extractor.frame_to_time(frame_indices=8), times[8])
 
         # Test image masks
         assert segmentation_extractor.get_roi_image_masks().shape == (self.num_rows, self.num_columns, self.num_rois)
@@ -107,3 +115,19 @@ class TestDummySegmentationExtractor(TestCase):
             exc_msg="'times' should have the same length of the number of frames!",
         ):
             segmentation_extractor.set_times(times_to_set)
+
+    def test_frame_to_time_no_sampling_frequency(self):
+        segmentation_extractor = generate_dummy_segmentation_extractor(
+            sampling_frequency=None,
+        )
+
+        times = np.arange(self.num_frames) / self.sampling_frequency
+        segmentation_extractor._times = times
+
+        self.assertEqual(segmentation_extractor.frame_to_time(frame_indices=2), times[2])
+        assert_array_equal(
+            segmentation_extractor.frame_to_time(
+                frame_indices=np.arange(self.num_frames),
+            ),
+            times,
+        )