+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 
 ### Improvements
+* Add `frame_to_time` to `SegmentationExtractor`, `get_roi_ids` is now a class method. [PR #187](https://github.com/catalystneuro/roiextractors/pull/187)
 
 ### Fixes